+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+- ?
+    + Added event handlers to apply the time range when 'Enter' is pressed in one of the time range inputs.
+
 - 0.2.1-1 (2025-01-24):
     + Improved rendering of timeseries.